+++ src/k8s/cmd/k8s/k8s_revoke_auth_token.go
@@ -1,0 +1,35 @@
+package k8s
+
+import (
+	"fmt"
+
+	"github.com/canonical/k8s/cmd/k8s/errors"
+	"github.com/spf13/cobra"
+)
+
+var (
+	revokeAuthTokenCmdOpts struct {
+		token string
+	}
+)
+
+func newRevokeAuthTokenCmd() *cobra.Command {
+	revokeAuthTokenCmd := &cobra.Command{
+		Use:               "revoke-auth-token --token <token>",
+		Short:             "Revoke an auth token for Kubernetes",
+		Hidden:            true,
+		PersistentPreRunE: chainPreRunHooks(hookSetupClient),
+		RunE: func(cmd *cobra.Command, args []string) (err error) {
+			defer errors.Transform(&err, nil)
+
+			err = k8sdClient.RevokeAuthToken(cmd.Context(), revokeAuthTokenCmdOpts.token)
+			if err != nil {
+				return fmt.Errorf("could not revoke auth token: %w", err)
+			}
+
+			return nil
+		},
+	}
+	revokeAuthTokenCmd.Flags().StringVar(&revokeAuthTokenCmdOpts.token, "token", "", "Token")
+	return revokeAuthTokenCmd
+}
