+++ src/k8s/pkg/k8s/client/kubernetes_auth_tokens.go
@@ -27,18 +27,17 @@ func (c *k8sdClient) GenerateAuthToken(ctx context.Context, username string, gro
 }
 
 // RevokeAuthToken calls "DELETE 1.0/kubernetes/auth/tokens".
-func (c *k8sdClient) RevokeAuthToken(ctx context.Context, token string) (string, error) {
+func (c *k8sdClient) RevokeAuthToken(ctx context.Context, token string) error {
 	queryCtx, cancel := context.WithTimeout(ctx, time.Second*30)
 	defer cancel()
 
 	request := apiv1.RevokeKubernetesAuthTokenRequest{Token: token}
-	response := apiv1.RevokeKubernetesAuthTokenResponse{}
 
-	err := c.Query(queryCtx, "DELETE", api.NewURL().Path("kubernetes", "auth", "tokens"), request, &response)
+	err := c.Query(queryCtx, "DELETE", api.NewURL().Path("kubernetes", "auth", "tokens"), request, nil)
 	if err != nil {
 		clientURL := c.mc.URL()
-		return "", fmt.Errorf("failed to query endpoint DELETE /kubernetes/auth/tokens on %q: %w", clientURL.String(), err)
+		return fmt.Errorf("failed to query endpoint DELETE /kubernetes/auth/tokens on %q: %w", clientURL.String(), err)
 	}
 
-	return response.Token, nil
+	return nil
 }
