+++ src/k8s/pkg/k8sd/api/impl/kubernetes_auth_tokens.go
@@ -22,12 +22,12 @@ func GetOrCreateAuthToken(ctx context.Context, state *state.State, username stri
 	return token, nil
 }
 
-func RevokeAuthToken(ctx context.Context, state *state.State, token string) (string, error) {
+func RevokeAuthToken(ctx context.Context, state *state.State, token string) error {
 	if err := state.Database.Transaction(ctx, func(ctx context.Context, tx *sql.Tx) error {
 		_, err := database.DeleteToken(ctx, tx, token)
 		return err
 	}); err != nil {
-		return "", fmt.Errorf("database transaction failed: %w", err)
+		return fmt.Errorf("database transaction failed: %w", err)
 	}
-	return token, nil
+	return nil
 }
